Various modifications to help provide good output to inferelator.
First of all, I added some new files to help label the data for inferelator.  In addition, added a file extractRWSD(extract regionals with standard deviation).  This program will allow you to get spread out regional triplicate data.

The file extractTriplits has been modified to provide a 3 dimensional array of data.  The third dimension can be though of as the experiment series number.  Also, the function returns the mean of the triplicates at a particular time point for a particular gene.  That is, the average of the values between experiments.